Move port collision logic to common app class
