+++ backend/src/main/java/ai/verta/modeldb/App.java
@@ -39,10 +39,7 @@ import io.grpc.BindableService;
 import io.grpc.ServerBuilder;
 import io.grpc.health.v1.HealthCheckResponse;
 import io.prometheus.client.Gauge;
-import java.io.BufferedReader;
-import java.io.File;
 import java.io.FileNotFoundException;
-import java.io.FileReader;
 import java.io.IOException;
 import java.sql.SQLException;
 import java.util.Optional;
@@ -57,7 +54,6 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.ApplicationPidFileWriter;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
@@ -254,33 +250,6 @@ public class App extends CommonApp {
     return false;
   }
 
-  public static void main(String[] args) throws Exception {
-    var path = System.getProperty(CommonConstants.USER_DIR) + "/" + ModelDBConstants.BACKEND_PID;
-    resolvePortCollisionIfExists(path);
-    SpringApplication application = new SpringApplication(App.class);
-    application.addListeners(new ApplicationPidFileWriter(path));
-    application.run(args);
-  }
-
-  private static void resolvePortCollisionIfExists(String path) throws Exception {
-    File pidFile = new File(path);
-    if (pidFile.exists()) {
-      try (BufferedReader reader = new BufferedReader(new FileReader(pidFile))) {
-        String pidString = reader.readLine();
-        var pid = Long.parseLong(pidString);
-        var process = ProcessHandle.of(pid);
-        if (process.isPresent()) {
-          var processHandle = process.get();
-          LOGGER.warn("Port is already used by modeldb_backend PID: {}", pid);
-          boolean destroyed = processHandle.destroy();
-          LOGGER.warn("Process kill completed `{}` for PID: {}", destroyed, pid);
-          processHandle = processHandle.onExit().get();
-          LOGGER.warn("Process is alive after kill: `{}`", processHandle.isAlive());
-        }
-      }
-    }
-  }
-
   @Bean
   public CommandLineRunner commandLineRunner(ApplicationContext ctx) {
     return args -> {
@@ -441,14 +410,5 @@ public class App extends CommonApp {
     initiateShutdown(0);
 
     cleanUpPIDFile();
-  }
-
-  private void cleanUpPIDFile() {
-    var path = System.getProperty(CommonConstants.USER_DIR) + "/" + ModelDBConstants.BACKEND_PID;
-    File pidFile = new File(path);
-    if (pidFile.exists()) {
-      pidFile.deleteOnExit();
-      LOGGER.trace(ModelDBConstants.BACKEND_PID + " file is deleted: {}", pidFile.exists());
-    }
   }
 }
